Prevent issuring tokens without TOTP finished.
The reason this bug exist is because the token is created with fullauthroity with appending acr with https://refeds.org/profile/mfa which if you open another window the authentication will retrieve the token from your session with only the PRE_AUTHENTICATED roles but going through else block since acr is not null anymore as if the upstream SAML already done TOTP. The IAM proceed to retrieve the full authrorities for the account and issue a new session with full authority. Since the token endpoint is not protected a token with user's full authriity is issued.

Thus this commit does the following:
- Append so that user baring SAML token session to go through TOTP authentication role instead of mistankenly taken as finished TOTP in upstream IdP.